Added propsal for a new editor interface.
Please note that this is intented as a base for an ongoing
discussion, not as final code.

I have not touched the rest of kdevelop in any way, so all
you get after make/make install (and installing the
simple-editor part) is a programm called "editortest".
But it is better to just read the headers. :)