Added full support for SuperWaba in kdevelop.  This is a pretty rough template and I have started to consider a real part that will handle some of the annoying things like downloading to the palm pilot and doing debugging.
The only part of this current system I really dislike is that any new
files must manually be added to the makefile.  I thought about a
Makefile editor, but my brain started to cry.

It is feasable that if I make a full SuperWaba part then I could use
automake, but I like the advantage here that the project can be
selfcontained and distributed as is.  Either way well see, this works
for what Im doing now so Im happy.  Feel free to kick it arround.