Fix some API doc errors. The warning messages that doxygen gave actually helped to clean up the code. I.e. qtbc.h was a file to enable backwords compatability with Qt1 and Qt2. We don't need it because we require Qt3.
It also fixes an error in kactionselector.h This file belongs to kate and it looks that not only KDevelop uses it but also kbabel. It migth be a good idea to move it to a common library.

Most of the remaining doxygen warnigs are "fixed" by using doxygen 1.3.5

CCMAIL: anders.lund@lund.tdcadsl.dk