This patch fixes: Bug 71003 : Cannot translate project due to srcdir != builddir. Bug 73467: Incomplete make clean for parts/appwizard/common
1. buildtools/autotools/addtranslationdlg.cpp
Problem here was that a make was called in <projectDir>/po, but when Build
configuration is debug there is no makefile in that directory.
So the make has to be called from <buildDir>/po.
The 'force-reedit' works fine because it changes internally to <projectDir>/po
for most of it's actual work.

2. parts/appwizard/common/kde-Makefile.am
This patch is slightly different from the patch proposed earlier.
'make-messages' has to work in the <projectDir> tree, taking the source from
<projectDir>/src and writing the .pot file to <projectDir>/po.

3. parts/appwizard/common/admin/am_edit
The '$(MAKE) -C po merge' form 'make messages' again needs to work in the
<projectDir>/po directory, merging the .pot file into the .po files.

4. parts/appwizard/common/Makefile.am
'make clean' does not remove the *.tar.gz files created in the parts/appwizard/common directory.

To make this patch work you need to:
cd parts/appwizard/common
make clean

Thanks to Frans Pop <aragorn  IN tiscali.nl> for the patch.

CCMAIL: 71003-done@bugs.kde.org
CCMAIL: 73467-done@bugs.kde.org