Add support for parsing classes and methods.
Heavily based on the python parsing.

Should work for non-complex code like:

class SomeClass < Superclass
	def someMethod
		...
	end

	def someOtherMethod
		...
	end
end

def anotherMethod
	...
end

But if the line with anotherMethod starts with anything except def
(like a whitespace) it will be reported as belonging to SomeClass,
which is a left-over from python, that I didn't bother to do anything
about right now, as it will work on most code, and it would require
a bit more thinking.