Okay this is not going into KDevelop because im not convinced this is the best approach anymore.  I have kicked this sucker arround since August, and it just wont get any better.
Ideally we can find a better direction for this to go.  Its not "visual" enough to please the group who wants
to nail together simple GUI apps... and its not small enough to work well for the group that wants to do
macros.

It sucks... but maby it can learn to suck less here.