Add in support for an additional license - GPL + Qt exemption. I thought about making this the default, but then realised it makes no sense on most KDevelop projects.
For those not familiar, the basic idea is that if you have GPL'd code,
but don't have a GPL'd Qt, you can't redistribute. This is a problem
for some commercial users who have Qt under other licenses, but the
main issue is that the Qt exemption allows you to use non-commercial
licensed Qt (ie the Windows version) to build GPL'd Qt apps.