Avoid bug 61031 in KDevelop 3.1
This converts '<', '>' and '&' to &gt; &lt; and &amp; if EscapeXML is true for a file in .kdevtemplate

Also removed handing a reference to the submap into all functions. This is available by a pointer in
the AppWizardDialog object.
Will fix templates now (basicly undo the install vs. installHTML change) to use this