There is some strange thing going on [1], as if I leave the KDevelop's statusbar as StatusBar I get the following error: /data/development/sources/kde-head/kdevelop/src/newmainwindow.cpp:323: error: syntax    error before `(' token
I verified and the "statusbar.h" header file is included and the double inclusion guards do not prevent the
definition of the StatusBar class.

The only solution was to rename StatusBar to KDevStatusBar.

[1] This happens on CVS HEAD, so I assume the introduction of setupGUI() and the StatusBar enum. I'm wondering if this can cause
crashes when running a KDevelop compiled for KDE 3.2.x under KDE 3.x or not. It's clear that KDevelop 3.0.x won't be compilable under KDE 3.3.


CCMAIL: kdevelop-devel@kdevelop.org