Fix a couple of bugs in the caption setting.
First, don't set a caption like "<projectname> - - KDevelop" when there aren't
any files open. I changed it to read "<projectname> - KDevelop" instead. When
there are files open it reads "<projectname> - <filename> - KDevelop" as before.

Also, fix a bug where the caption wasn't being set correctly when the project was
loaded. The wrong caption is still set when unloading a project, so that'll need to be
fixed later, but this is a start in the right direction.