backport revision 1.19.
This fix should be in KDevelop 3.2.1
CCBUGS: 99590


Original log message:
If the line already contains tags (or any other text surrounded by "<" and ">")
we need to replace these delimiters with their corresponding HTML code.
That way the tags will be no longer recognized by QTextEdit but displayed
as plain text.
Besides that QTextEdit crashes in some special constellations if we
try to wrap or own tags around already existing ones.