Properly resize first column of variables view.
The existing code tries to set var name column width to be half the total
width, but does not succeed. The width is set in VariableWidget::setEnabled,
which is called in thise context.

   #0  GDBDebugger::VariableWidget::setEnabled (this=0x817a5a8, bEnabled=true) at variablewidget.cpp:127
   ...
   #3  0xb742013d in KDockWidget::setWidget () from /usr/lib/libkdeui.so.4
   #4  0xb7dbdff5 in KMdiToolViewAccessor::KMdiToolViewAccessor () from /usr/lib/libkmdi.so.1
   #5  0xb7dabe47 in KMdiMainFrm::addToolWindow () from /usr/lib/libkmdi.so.1

Here, the width of variable view is still the same as was at creation time
(640 for me) and after KMDI sets the the real width of the vartree, the column
width becomes was too large.

This commit resizes the column on the first showEvent.

CC: kdevelop-devel@kdevelop.org