Ubreak display of QStrings broken in r436822.
* variablewidget.cpp
  (VarItem::checkForRequest): Rename to 'handleSpecialTypes'. Instead
  of looking at data, check the type of variable.
  (VarItem::updateValue): Remove \000 from gdb output.

Below is the exact explanation of the breakage. 

The previous code worked like this for the case of reference to QString
1. Gdb sends something like s = {d = 0x1234}
2. Parser create VarItem and calls setCache("{d = 0x1234}")
3. setCache does not recognize anything yet (due to leading "{")
4. User clicks on the "open"
5. VarItem::setOpen is called, it calls gdb parser again with "{d = 0x1234}"
6. Parser notices it's a structure so it stripts "{} and calls setCache
   again.
7. setCache notices "d = 0x1234" and asks for string value.
8. setCache calls setOpen again, which calls gdb parser again, with
   "d = 0x1234"
9. Parser sets the 'd' child of variable.
10. The string value arrives.

After my changes in 436822 the double call to parser (step 5 and 8) are gone,
now gdb parser strips {} immediately and proceeds to setting the 'd' child,
so the code that used to detect QString never worked.

It was kind of a hack anyway, and it detected 'd = 0x' pattern, and not the
type of variable.

CCMAIL: amantia@kde.org