Prevent debugger from becoming unusable if gdb crashes.
When debugger issues any command, it wraps it in two markers, say:

   set prompt \032V
   set variable foo = bar

causes gdb to reply with
  \032V <output from the command> \032V

This allows the debugger to find output that correspond to any given command.
In the above example, if "set variable" command crashes the debugger, the
internal buffer for gdb output will contains just "\032V". That buffer was not
cleared, so after restart KDevelop would try to find the second \032V,
and never succeed. It looked like debugger is running, but user could not
enter any commands and all debugger actions did not work.

CCBUG: 97795