Fix sending bogus "condition" command to gdb.
The problem occurred when changing any property of breakpoint except for
condition. On initialization, the condition_ field of breakpoint was set to
"". When parsing breakpoint list the local QString variable had no
initializers. When the local variable was passed to
Breakpoint::setConditional, and comparison between QString("") and QString()
was made, result was false, and the s_changedConditional_ was set to true.
This causes "condition N %s" to be sent, where "%s" was empty QString(), and
so became "(null)" literal, and that confused gdb.

The real problem here is that that 'conditional changed' flag should only be
set when the value is edited by the user, not when it's changed as result of
pulling data from gdb. But that's for later.