Add KDevConfigSkeleton class that will handle writing to the local project file in the hidden directory.
It'll cycle through the items and for 
those items marked as 'nonshareable' it'll force the item to write to the 
hidden directory file instead.  And we'll mark items by reading in a 
local.kdev4 file that is installed with each .kfcg file that provides a 
simple list of the nonshareable setting names.