Add support for CMake to KDevelop.
What this does is allows one to view the sources and targets as CMake sees them
in the project manager. It doesn't build anything (yet)

There are a few things that are required in order for this to work correctly:

1. You must point the environment variable "CMAKE_ROOT" to the location of the
CMake data directories. For example, if you installed CMake with a prefix of
/usr, you should do 'export CMAKE_ROOT=/usr/share/CMake'. This is so that CMake
can find all it's stuff correctly and actually return the information on targets,
sources, etc.

2. You need to add something like the following to your project file you use in
kdevelop4 or you risk the possibility of messing up your build environment:

  <cmakeproject>
    <builddir>/home/kde/trunk/build/KDE/kdevelop</builddir>
    <prefix>/home/kde/installs/4.0</prefix>
    <buildtype>debugfull</buildtype>
  </cmakeproject>

The builddir setting is key here. Since the CMakeCache.txt file will be loaded (and
be used for other cool things later). The other two settings are currently ignored
since we don't actually support compiling anything yet. If you want to make sure
you don't mess it up, find me on IRC or send me and email, and I'll send you an
example file with these settings in it for you to tweak and customize.

Unfortunately, I had to add nearly all of cmake to kdevelop's source tree. I don't
like doing this. This is what the cmake folks told me to do however,  and I hope to
be doing the following things in the future to improve the situation:

1. Allow CMake to depend on external versions of expat and curl and use those instead
of the built in copies that CMake ships with itself.

2. Remove a bunch of the extraneous code that's not and will probably never be used.
The code for all the various commands comes to mind here. 

3. Work with the CMake folks so that CMake exports a nice API that we can use to read
the files (and possibly write them back)

Now the fun starts. :)

CCMAIL: kdevelop-devel@kdevelop.org