And we're parsing java files... :)
All you need to start parsing java files is to change
the line PrimaryLanguage=C++ in the global project file
to PrimaryLanguage=Java.

Right now the codemodel is not hooked up and the AST is 
not integrated.

I don't think CMake knows how to deal with .ll files, so
add the .cc file instead.

The kdevelop-pg loads the _G_contents char array in a weird way.  Need
to change this so that it uses QByteArray or however the c++ parser
does it.

I commented out a few lines in java-io.cpp that were referencing main.cpp
these methods in main.cpp need to be moved into a separate file if we
want to keep them.

Next steps would be to integrate the AST so that it inherits KDevAST and then
we need to generate an actual codemodel that inherits KDevCodeModel.  Once
this is done, we'll be able to see the codemodel in the codeview :)

Cheers,Adam

CCMAIL:jpetso@gmx.at, kdevelop-devel@kdevelop.org