Yet Another Patch from Andreas Pakulat.
"the attached patch add's support for subclassing Qt4 forms. It uses the
multiple inheritance approach via a new template. I figured this is the
most easiest way to provide support for Ui:: namespace, ui_ header
naming and the removal of the name and model attributes from the
constructor.

A nice "improvement" in case of mixed qt3 and qt4 .ui files would be to
read the version of the ui-file. But I think I can leave that for later,
as probably only very few people have such a mixed setup within the same
project..."

Thanks for the patch!
CCMAIL: apaku@gmx.de