KDevCore provides a set of fundamental objects for the KDevPlatform.
These objects sometimes depend upon each other, but we can not know
how an application will order the creation of the various objects.

Because of this, all initialization of resources that depend upon one
or more KDevCore objects should go in an 'initialize' method that is
called by KDevCore::initialize();  Similarly, all cleanup of resources
that depend upon one or more KDevCore objects should go in a 'cleanUp'
method that is called by KDevMainWindow::queryClose();  Once this
cleanUp is performed the objects can be safely deleted from KDevCore.

CCMAIL: kdevelop-devel@kdevelop.org