Add a stub for KDevPeristentHash.  It is just a stub for now because it requires google's sparsehash.  It works, but I don't want to force everyone to install sparsehash until it is useful.
KDevAST is now a struct and has a void* pointer for storing
the associated memory pool.  We'll probably want to store
the memory pool in another way at some point

Add some stub methods in KDevLanguagePart and the csharp
language part for using the new serialize AST.  An impl
is provided in KDevLanguagePart for now so that C++ doesn't
have to provide it.

Provide actions for opening and closing projects.