Major rethink of definitions occurring in function parameters and if statements.
Now utilises the direct acyclic graph design (suggested by Roberto),
rather than being a plain tree. If this was what Roberto meant to start
with, I didn't understand it until now ;)  This is one of those commits
that makes you feel content that you've done it the "right way".

So, the DUChain is looking much better. Perhaps now it's time to work
on getting #include-ing of ASTs, or incremental parsing, I haven't
decided yet...