Major speedup brought about by removing QTextStream.
I knew that QTextStream would be a bit of a compromise, but I had no
idea how much it actually was.

This shaves 50% off the time to fully parse a file, so the old code must
be at least an order of magnitude worse.

Anyway, in total I've taken 80% off the time to parse a file... woohoo!