Semi-working definition use chain viewer.  Seems the model doesn't like dynamic_casting when the object is not the first class to be inherited?
The model works but it infinitely recurses on the last item in the chain
at each level (bug) so i limited it to 20 items per level.

To experiment, uncomment the connect line in DUChainModel::DUChainModel.

Warning: if you trigger reparsing (eg. type something) then fiddle with
the viewer, it will crash.  To be fixed :)