Last night and tonight's work...:
1) Use static deleters for singletons

2) Use new KUrl for KDevDocumentRangeObject when working with non-smart
cursors.  I found that even a blank KUrl (QUrl) allocates a QUrlPrivate,
and this was hurting us - 5Mb memory use on a 1Mb document.  Found on
profiling with Massif.

3) Locking for the definition-use chain.
   - One top-level lock per chain for most functions
   - Per-object locking (internally) for external references - not
     finished
   - Use locking in client code

I have left debugging enabled for now... if you hit an assert, please
email/irc me.

4) Disable the duchainviewer.  For this to work, there is still a fair
bit of work to be done:
   - need notification of major changes to a duchain while unlocked
   - need to create a proxy model which can forward data from any thread
     to the gui thread