BUG: 111686 BUG: 67746
Importing project - version is read from configure.[in.in|in|ac]. Check for configure.[in|in.in|ac] when trying to create automake project, and don't fail if there is no AUTHOR file. 
Changing the version in project options - updates configure.[in.in|in|ac] if they exist. Added tool tip advising that automake may need to be run to update all files.
Removed hard coded references to configure.in - if imported project uses configure.ac, then this will be updated.
Tooltip in import dialog telling format of author/email fields.