Fix support for $$FOOBAR/somefile.cpp and similar, we can now open them upon clicking in the Ui and add such things.
This adds a new QMake Manager configuration option which controls wether relative paths of files are replaced by variables from the .pro file (if a variable containing the exact relative path exists). So Steven as you want that open up the Project Options and look at the QMake Manager Options.

Steven if you find that this broke anything, please re-open this bug.

BUG: 137768