Decouple the database backends for the catalog from the catalog itself.
Introduce an abstract class called KDevCatalogBackend that all the calls to a 
database are proxied through. Most functions in KDevCatalog now call functions
in KDevCatalogBackend.

An initial implementation for bdb has been provided since that was what was in
use before.

Still TODO: write checks for bdb to the build system, add it to the build system
and make it compile.