Move the C# language part into kdevelop4-extra-plugins.
It features a fully-functional parser and
a binder that hooks up with KDevelop's codemodel,
but the binder doesn't have such a great future,
it should be redone for better integration with
DUChains (which are not yet existing at all
in the C# part).

The best thing would be to drop the binder
altogether, but this still requires improvements
in DUChain itself.