KDevelop say hello to your new way of handling context menus.
This is a first simple implementation, without the weights or any other extra
logic.  I added a build entry to the project manager view context menu to
demonstrate it.  The docs on the Context class got updated with some hints what
a plugin has to do.

Unfortunately for some unknown reason the signal mapper doesn't work, so only
the dummy buildSelectedItem slot is called at the moment. If anybody can give
me a hint whats wrong there that would be great

CCMAIL: kdevelop-devel@kdevelop.org