Redo the move, this time the history is preserved, this also means the changes look slightly different than the first time( some deletes some moves).
!!!!!! IMPORTANT NOTE !!!!!!!!!!!
Everybody who wants to change something in kdevelop/lib or one of the plugins
that are copied to the platform (snippet, teamwork, konsole, filemanager,
duchainviewer):
Don't, checkout kdevplatform and do the change there. I'm starting to port
kdevelop to use the platform tomorrow and expect to be done on wednesday
evening. I will do this in a separate branch and the first thing that will
happen are deletion of lib/ and the mentioned plugins. So any changes you do
between now and the time I merge the branch back may get lost.

CCMAIL:kdevelop-devel@kdevelop.org