Allow to easily retrieve the exact binary used in a project. This will be used by the QMake Manager.
CC'ing kdevelop-devel to get some input on a question wrt. to this change:
Would it be better to offer something like
QString queryQMake(const QString& var, KDevelop::IProject*)
which would basically run qmake -query var and return the result?

I need a way in the qmake manager to run qmake -query var but am not 100%
sure which of the two API's is "better" and I'd like to not hardcode the qmake group
config-key in two places..

CCMAIL:kdevelop-devel@kdevelop.org