Cool Feature Of The Day(TM)
Move calculation of code completions into a separate thread, thus not blocking the ui (as much) when it is invoked.
To me, this makes automatic code completion invocation more bearable, but still not ready for daily use.

Note: you need an up-to-date katepart in order to see completions now, due to the bugs I fixed in katepart today.