Implement dock widget resizing with a custom splitter handle. Make dock widget anchoring behave like kdev3.
There is still a bug when you resize the bottom dock:
the layout is continually invalidated by QEvent::LayoutRequest events.

I spent an hour trying to figure out the source of this loop, but so far
have failed.  It only seems to happen on some of the dock widgets, so it
 must be related to the dock wigets' contents.

A big cookie to anyone who can figure it out :)