Implement moving of tool view between window sides.
Presently, this is implemented as popup menu on tool's
title. Neither this solution, nor implementation is ideal.
At the very least, we need to be also able to move
by context menu on the button that selects the tool. Ideally,
we should be able to drag-drop tools, and make two or several
tools simultanously present on one side of the screen.
	
	mainwindow_p.h, mainwindow_p.cpp/
	* MainWindowPrivate (toolViewMoved): New slot.	
	area.h, area.cpp/
	* Area (moveToolView): New method.
	ideal.h, ideal.cpp/
	* IdealDockWidgetTitle (IdealDockWidgetTitle): New 
	parameter docking_area. 
	(contextMenuEvent): New.
	* IdealButtonBarWidget (addWidget): 
	Pass docking position to IdealDockWidgetTitle.
	* IdealMainWidget (addView): If view returns
	a parentless widget, set parent explicitly.
	(removeView): New parameter 'nondestructive'.
	Reset view parent if specified.
	(moveView): New.