Fix crash when closing multiple windows.
If a file is shown in two windows, we have one 
KTextEditor::Document with several views, and we
actually try to delete views ourself. However, by
default if a view for KParts::Part dies, the 
part itself dies too (closing othe views). As result,
we end up with already-deleted view that we try to
delete.