Change our SOVERSION to 1 as discussed. Claiming these libs had already 4 BiC releases is just false.
Sorry for being a bit late, I totally forgot about this on thursday.

Note for anybody with an existing kdevplatform build: After this update you
first need to completely remove any libkdevplatform* in your builddir and
installation dir and then start the build.
AFAIK no reason to completely remove the builddir for kdevplatform or kdevelop,
just the above and then run a simple make in both.

CCMAIL:kdevelop-devel