Fix xmlgui crashes when closing views.
We failed to remove XMLGUIClient for Kate views when closing those views.
As result KXMLGUIFactory used to hold on to deleted clients, and when
we eventually created new view and tried to switch to it, "removing"
old, already-deleted, one, things would crash. Or not crash, depending
on luck, but usually, 10 attempts at opening/closing view were sufficient
for segfault to appear.