Allow to close tabs by clicking the "close" button.
Along the way, discovered that if one side of split view is all closed,
we actually fail to move all content of the other side to the parent. This
was not noticable because QSplitter with 1 child does not have any decoration,
but apparently was not intended, so I've fixed that.