Disable the DEBUG_UI_LOCKUP define by default.
All the timer stuff in there is murder for those of us who spend a lot of time
working off of our laptop's battery.

Eventually, we can make this define be set by the buildsystem so a simple
"make edit_cache" will allow developers to turn it on and off easily.