- When creating duchain contexts from within macros, like for example within the foreach macro, always create then at the end of the macro use, with a collapsed range. - If there's a declaration within such a collapsed context that comes from a valid text position, like the first foreach(..) argument, give that declaration a parent-range that encloses its range. - Add some optional debugging functionalities
In order to achieve the first point, the preprocessor has to track with each marked position whether it has been marked from within a macro expansion.

The immediate result of all this work: Code that contains foreach(..) statements/macros is updated correctly, and the highlighting doesn't flash while parsing there.