Add the new context browser. This contains the previous use-highlighter.
What it does: It currently contains 2 sections in the tool view:
One showing you information about the context you're in using a html navigation widget, and allowing you jumping to previous and next used contexts(including keyboard shortcuts for that).
And the other section showing you information about the declaration you're currently dealing with, and that is also use-highlighted. This now also works with the mouse, sou you can just point at a declaration or use, and you'll see all the information about it in the context browser.