Large refactoring to allow running state of plugins to be shown to the user, and jobs to be killed...:
Change IRunController to allow it to manage KJobs.  You can now register (and start) a new job by calling registerJob().

Change IRunProvider to work with KJobs instead of serial numbers (much better design :)

Add a menu option for selecting which project should be run, so the run configuration works again (but not the add target feature, will work on that later)

Created new KJob subclass that handles the creation of an output view to correspond to that KJob (in the utils module, OutputJob)

Added feature to output interface/plugin to allow the plugin to take ownership of a model (so that jobs which are deleted can still show their output)

Port to KJob
- cvs plugin
- svn plugin (could only find one place that jobs are started, is that correct??)
- execute (run) plugin