Obey KDE wide single/double click setting.
This poses a bit of a problem though, its now kind of hard to only select one file in the tree without opening it. The workaround the dolphin uses isn't going to work as we have to few space.

OTOH forcing double-click is also bad because then we need to disable editing of items via double clicking as it gets in the way otherwise.

Anybody with bright ideas how to solve this dilemma?

CCMAIL:kdevelop-devel@kdevelop.org