We must iterate all areas (including clones) to completely remove tool views.
This fixes the crash shortly after closing all projects, because without it some areas still have their reference to the now deleted view.

However, this reveals a new problem - when we open a new project, the project tool view is not automatically re-added, you have to add it via view->add tool view... shouldn't this be automatic?

CC:kdevelop-devel@kdevelop.org