Use the AbstractDeclarationBuilder
Not much of the generic stuff can be used, because openDeclaration is a huge, complex method with lots of template stuff... maybe it should be refactored at some point.  For now just keep the c++ specific version... anyone know how to tell 
GCC we want the base class functions to be hidden? (it gives off warnings)