Take a different approach to document loading: tell the smart interface not to delete smart ranges on reload.  Then we can just reposition them on the next run. Document reloading now crashes even less often (I haven't seen it crash yet). However it still doesn't re-highlight after reload, more work to do there.
Change the locking in SmartConverter - it now requires the duchain read-write lock to be held before calling, so it is safe to get the smart lock.  I added the generic code back to the context builder ready for use, but i'm still not 
convinced it's needed (more investigation required).

Some apidox updates (my doxygen skills are rusty, how do you build it nicely like on api.kde.org?)