- Assign an index to each top-context, make each top-context know that index, and allow retrieving declarations from the du-chain based on that index. - To each Declaration, assign an index within its surrounding top-context.
Together, this allows globally adressing declarations using these two indices. Currently it adds overhead, but only until the transition to a serializable structure is complete.
These indices will be used in DeclarationId, which will address declarations that are in the symbol-table using the qualified identifier, and declarations that are not within the symbol-table using these indices.