- Change ExpressionEvaluationResult so it can store itself in a repository, and add IndexedExpressionEvaluationResult that uses that repository - Add a new repository-managed class InstantiationInformation that stores a complete chain of applied instantiations, and that can be used to get any instantiated template from its base template, independent of the parent context. - Change the complete template instantiation system in the way it works, so any template instantiation can be retrieved from its basic declaration plus the InstantiationInformation, without any further context information. Together with the storage of the instantiation information in a repository, and adding that index to DeclarationId and using it from there, this means that now we can get any template instantiation at runtime indirectly, just using qualified identifiers or direct declaration indices, plus the specialization information index.
We now have a clear separation between the instantiations, which can be computed on first access and just deleted any time, and the declarations that are directly 
related to the code. We can just save those to disk, and all instantiations will still be accessible using their DeclarationId.

All the tests pass again, and it also doesn't seem to crash any more, but there's temporary speed regressions.