Hopefully fix >99% of the crashes which can happen when you close a document, especially while the document is being duchain-parsed.
The editor integrator no longer returns the SmartInterface directly, now it returns an object (LockedSmartInterface) which holds the interface in a locked state while it is needed.  This should eliminate the race to acquire the smart lock 
at the same time as the document is being closed.  Users of the editor integrator need only slightly update their code.