Make the whole repository storing actually work, and add a lot of optional debugging code to itemrepository.h.
Pick the storage-directory from within itemrepository.cpp instead of duchain.cpp.

It is important that the repositories need to be loaded at the earlierst point possible. The problem: At that early point, neither KStandardDirs, nor KLockFile are functioncal, because KComponentData doesn't exist yet.
So for now we just store the data in ~/.kdevduchain, and don't support multiple running instances of kdevelop until we find another solution for the locking problem.