Adding / comitting changes to files not yet added to the build system.
I was wrong earlier, c++ makes extensive use of ConstantIntegralType seperate from EnumeratorType, so separate them again.

The full patch looks like it's going to take a while to get right, and with the big changes happening in duchain persistance, I've decided to put them on the backburner for 
the moment.