Finally expose accurate document loaded signals for text documents (and others, when they would like to use it).
To do this, I added a new signal to IDocumentController - documentOpened().  This takes the place of the old documentLoaded(), in that the document controller emits it when 
the document has been created.  The new documentLoaded() signal is emitted via the document itself (by calling IDocument::notifyLoaded()), when the file has been loaded 
from disk/network etc.  This means the documentLoaded() may be delayed a while after documentOpened(), or even come before documentOpened(), but it will always be when the 
file has been loaded from disk.  This should solve the problem that contexts which are loaded from disk are not highlighted. (although I can't test this, because kdevelop 
always crashes on close and even removing the lock file doesn't seem to let it load previously parsed repositories, or they are not saved because of the crash - next step 
is to work on the crash(es))