Fix crashes related to stale smart ranges in editor integrators when files are closed.
To fix this, I audited (again!) the editor integrator class, and now you have to pass the locked smart interface to any function which works with the smart interface 
or smart ranges.  This will probably result in a performance improvement in cases where the editor integrator tried for the lock itself, because the lock was usually 
already held.

Also change the abstract declaration builder's openDeclaration/openDefinition functions to be template-based like in c++, so you can create custom declaration classes 
and use them with the generic function.